implement IP allocation for internal LoadBalancer

+++ apinetlet/controllers/loadbalancer_controller.go
@@ -16,17 +16,12 @@ package controllers
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"net/netip"
 	"strings"
 
 	"github.com/go-logr/logr"
-	"github.com/onmetal/controller-utils/clientutils"
-	onmetalapinetv1alpha1 "github.com/onmetal/onmetal-api-net/api/v1alpha1"
-	apinetletv1alpha1 "github.com/onmetal/onmetal-api-net/apinetlet/api/v1alpha1"
-	commonv1alpha1 "github.com/onmetal/onmetal-api/api/common/v1alpha1"
-	networkingv1alpha1 "github.com/onmetal/onmetal-api/api/networking/v1alpha1"
-	"github.com/onmetal/onmetal-api/utils/predicates"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -36,8 +31,16 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/cluster"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
-	"sigs.k8s.io/controller-runtime/pkg/predicate"
 	"sigs.k8s.io/controller-runtime/pkg/source"
+
+	"github.com/onmetal/controller-utils/clientutils"
+	onmetalapinetv1alpha1 "github.com/onmetal/onmetal-api-net/api/v1alpha1"
+	apinetletv1alpha1 "github.com/onmetal/onmetal-api-net/apinetlet/api/v1alpha1"
+	commonv1alpha1 "github.com/onmetal/onmetal-api/api/common/v1alpha1"
+	ipamv1alpha1 "github.com/onmetal/onmetal-api/api/ipam/v1alpha1"
+	networkingv1alpha1 "github.com/onmetal/onmetal-api/api/networking/v1alpha1"
+	client2 "github.com/onmetal/onmetal-api/utils/client"
+	"github.com/onmetal/onmetal-api/utils/predicates"
 )
 
 const (
@@ -88,7 +91,18 @@ func (r *LoadBalancerReconciler) deleteGone(ctx context.Context, log logr.Logger
 		return ctrl.Result{}, fmt.Errorf("error deleting apinet public ips: %w", err)
 	}
 
-	log.V(1).Info("Issued delete for any leftover apinet public ip")
+	log.V(1).Info("Deleting any matching apinet prefix")
+	if err := r.APINetClient.DeleteAllOf(ctx, &ipamv1alpha1.Prefix{},
+		client.InNamespace(r.APINetNamespace),
+		client.MatchingLabels{
+			apinetletv1alpha1.LoadBalancerNamespaceLabel: virtualIPKey.Namespace,
+			apinetletv1alpha1.LoadBalancerNameLabel:      virtualIPKey.Name,
+		},
+	); err != nil {
+		return ctrl.Result{}, fmt.Errorf("error deleting apinet prefix: %w", err)
+	}
+
+	log.V(1).Info("Issued delete for any leftover apinet public ip / prefix")
 	return ctrl.Result{}, nil
 }
 
@@ -108,28 +122,43 @@ func (r *LoadBalancerReconciler) delete(ctx context.Context, log logr.Logger, lo
 		return ctrl.Result{}, nil
 	}
 
-	var count int
-	for _, ipFamily := range loadBalancer.Spec.IPFamilies {
-		if err := r.APINetClient.Delete(ctx, &onmetalapinetv1alpha1.PublicIP{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace: r.APINetNamespace,
-				Name:      fmt.Sprintf("%s-%s", loadBalancer.UID, strings.ToLower(string(ipFamily))),
-			},
-		}); err != nil {
-			if !apierrors.IsNotFound(err) {
-				return ctrl.Result{}, fmt.Errorf("error deleting target public ip: %w", err)
+	if loadBalancer.Spec.Type == networkingv1alpha1.LoadBalancerTypeInternal {
+		for _, ipSource := range loadBalancer.Spec.IPs {
+			if err := r.APINetClient.Delete(ctx, &ipamv1alpha1.Prefix{
+				ObjectMeta: metav1.ObjectMeta{
+					Namespace: r.APINetNamespace,
+					Name:      fmt.Sprintf("%s-%s", loadBalancer.UID, strings.ToLower(string(ipSource.Ephemeral.PrefixTemplate.Spec.IPFamily))),
+				},
+			}); err != nil {
+				if !apierrors.IsNotFound(err) {
+					return ctrl.Result{}, fmt.Errorf("error deleting prefix: %w", err)
+				}
 			}
-			count++
+		}
+	} else {
+		var count int
+		for _, ipFamily := range loadBalancer.Spec.IPFamilies {
+			if err := r.APINetClient.Delete(ctx, &onmetalapinetv1alpha1.PublicIP{
+				ObjectMeta: metav1.ObjectMeta{
+					Namespace: r.APINetNamespace,
+					Name:      fmt.Sprintf("%s-%s", loadBalancer.UID, strings.ToLower(string(ipFamily))),
+				},
+			}); err != nil {
+				if !apierrors.IsNotFound(err) {
+					return ctrl.Result{}, fmt.Errorf("error deleting target public ip: %w", err)
+				}
+				count++
 
+			}
+		}
+
+		if count < len(loadBalancer.Spec.IPFamilies) {
+			log.V(1).Info("Target public ip is not yet gone, requeueing")
+			return ctrl.Result{Requeue: true}, nil
 		}
 	}
 
-	if count < len(loadBalancer.Spec.IPFamilies) {
-		log.V(1).Info("Target public ip is not yet gone, requeueing")
-		return ctrl.Result{Requeue: true}, nil
-	}
-
-	log.V(1).Info("Target public ip is gone, removing finalizer")
+	log.V(1).Info("Target public ip / prefix is gone, removing finalizer")
 	if err := clientutils.PatchRemoveFinalizer(ctx, r.Client, loadBalancer, loadBalancerFinalizer); err != nil {
 		return ctrl.Result{}, fmt.Errorf("error removing finalizer: %w", err)
 	}
@@ -150,11 +179,18 @@ func (r *LoadBalancerReconciler) reconcile(ctx context.Context, log logr.Logger,
 		return ctrl.Result{Requeue: true}, nil
 	}
 
-	ips, err := r.applyPublicIPs(ctx, log, loadBalancer)
-	if err != nil {
-		return ctrl.Result{}, fmt.Errorf("error getting / applying public ip: %w", err)
+	var ips []netip.Addr
+	if loadBalancer.Spec.Type == networkingv1alpha1.LoadBalancerTypeInternal {
+		ips, err = r.applyPrivateIPs(ctx, log, loadBalancer)
+		if err != nil {
+			return ctrl.Result{}, fmt.Errorf("error getting / applying private ip: %w", err)
+		}
+	} else {
+		ips, err = r.applyPublicIPs(ctx, log, loadBalancer)
+		if err != nil {
+			return ctrl.Result{}, fmt.Errorf("error getting / applying public ip: %w", err)
+		}
 	}
-
 	if err := r.patchStatus(ctx, log, loadBalancer, ips); err != nil {
 		return ctrl.Result{}, fmt.Errorf("error patching load balancer status")
 	}
@@ -212,6 +248,61 @@ func (r *LoadBalancerReconciler) applyPublicIP(ctx context.Context, log logr.Log
 	return ip.Addr, nil
 }
 
+func (r *LoadBalancerReconciler) applyPrivateIPs(ctx context.Context, log logr.Logger, loadBalancer *networkingv1alpha1.LoadBalancer) ([]netip.Addr, error) {
+	var ips []netip.Addr
+	for idx, ipSource := range loadBalancer.Spec.IPs {
+		apiNetPrivateIP, err := r.applyPrivateIP(ctx, log, loadBalancer, ipSource, idx)
+		if err != nil {
+			return nil, fmt.Errorf("[ip %d] %w", idx, err)
+		}
+
+		ips = append(ips, apiNetPrivateIP)
+	}
+	return ips, nil
+}
+
+func (r *LoadBalancerReconciler) applyPrivateIP(ctx context.Context, log logr.Logger, loadBalancer *networkingv1alpha1.LoadBalancer, ipSource networkingv1alpha1.IPSource, idx int) (netip.Addr, error) {
+	switch {
+	case ipSource.Value != nil:
+		return ipSource.Value.Addr, nil
+	case ipSource.Ephemeral != nil:
+		template := ipSource.Ephemeral.PrefixTemplate
+		prefix := &ipamv1alpha1.Prefix{
+			ObjectMeta: metav1.ObjectMeta{
+				Namespace: r.APINetNamespace,
+				Name:      fmt.Sprintf("%s-%s", loadBalancer.UID, strings.ToLower(string(template.Spec.IPFamily))),
+			},
+		}
+		log.V(1).Info("Applying prefix", "ipFamily", template.Spec.IPFamily)
+		if err := client2.ControlledCreateOrGet(ctx, r.Client, loadBalancer, prefix, func() error {
+			prefix.Labels = template.Labels
+			if prefix.Labels == nil {
+				prefix.Labels = make(map[string]string)
+			}
+			prefix.Labels[apinetletv1alpha1.LoadBalancerNamespaceLabel] = loadBalancer.Namespace
+			prefix.Labels[apinetletv1alpha1.LoadBalancerNameLabel] = loadBalancer.Name
+			prefix.Labels[apinetletv1alpha1.LoadBalancerUIDLabel] = string(loadBalancer.UID)
+			prefix.Annotations = template.Annotations
+			prefix.Spec = template.Spec
+			return nil
+		}); err != nil {
+			if !errors.Is(err, client2.ErrNotControlled) {
+				return netip.Addr{}, fmt.Errorf("error managing ephemeral prefix %s: %w", prefix.Name, err)
+			}
+			return netip.Addr{}, fmt.Errorf("prefix %s cannot be managed", prefix.Name)
+		}
+
+		if prefix.Status.Phase != ipamv1alpha1.PrefixPhaseAllocated {
+			return netip.Addr{}, fmt.Errorf("prefix %s is not in state %s but %s", prefix.Name, ipamv1alpha1.PrefixPhaseAllocated, prefix.Status.Phase)
+		}
+		log.V(1).Info("Retuen prefix ip")
+
+		return prefix.Spec.Prefix.IP().Addr, nil
+	default:
+		return netip.Addr{}, fmt.Errorf("unknown ip source %#v", ipSource)
+	}
+}
+
 func (r *LoadBalancerReconciler) patchStatus(ctx context.Context, log logr.Logger, loadBalancer *networkingv1alpha1.LoadBalancer, ips []netip.Addr) error {
 	base := loadBalancer.DeepCopy()
 	loadBalancer.Status.IPs = []commonv1alpha1.IP{}
@@ -231,11 +322,6 @@ func (r *LoadBalancerReconciler) patchStatus(ctx context.Context, log logr.Logge
 	return r.Status().Patch(ctx, loadBalancer, client.MergeFrom(base))
 }
 
-var isLoadBalancerTypePublic = predicate.NewPredicateFuncs(func(obj client.Object) bool {
-	loadBalancer := obj.(*networkingv1alpha1.LoadBalancer)
-	return loadBalancer.Spec.Type == networkingv1alpha1.LoadBalancerTypePublic
-})
-
 func (r *LoadBalancerReconciler) SetupWithManager(mgr ctrl.Manager, apiNetCluster cluster.Cluster) error {
 	log := ctrl.Log.WithName("loadbalancer").WithName("setup")
 
@@ -245,7 +331,6 @@ func (r *LoadBalancerReconciler) SetupWithManager(mgr ctrl.Manager, apiNetCluste
 			builder.WithPredicates(
 				predicates.ResourceHasFilterLabel(log, r.WatchFilterValue),
 				predicates.ResourceIsNotExternallyManaged(log),
-				isLoadBalancerTypePublic,
 			),
 		).
 		Watches(
